test(e2e): Fix node E2E test app
